Add files via upload
I initially got a graph in which my strong prior and sample of 10 were similar when graphed. I know that was expected because the small sample size doesn't overwhelm our prior.  However, when I ran the code again, after R forced me to restart, the graph changed and I didn't change the code so I don't know why the graph looks off now but my strong prior should be on the right side of the graph, aligning with the blue curve because that is the sample 10 and that is how it looked originally. I just wanted to let you know that my explanation is written out but my graph just won't show it even though I tried knitting it 4 separate times.